+++ GameProject/Game/GameLogic/ReplaySystem.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Engine/Events/Events.h>
+#include <Game/GameLogic/GameEvents.h>
 #include <Game/Components/PathTreader.h>
 #include <Utils/Timer.h>
 #include <vector>
